requirements.txt
flask>=3.0.0
yt-dlp>=2025.12.8
gunicorn>=21.2.0
python-dotenv>=1.0.0
